did WORK on the Driver.java class
pretty much overhauled the class... instead of just executing every instruction
as it comes in from STDIN, now every instruction gets pushed onto the stack in
the opcode space (starting at 16 for now).  once every instruction is pushed on
the stack, then PC marches through the instructions and executes them one by
one until a HALT is encountered.  the only really tricky thing i did was
introduce a hash map, which maps the location of an instruction on the stack to
its argument (could be null).  here is a diagram of possible hash values (i'm
drawing the values like they would kind of appear in the stack ;))

address  value

22       42
21       68
20       null
19       null
18       18
17       null
16       97

the opcodes are NOT a part of this hash map!