got `fib.sxx' producing correct output!
the problem was twofold; i was subtracting 1 from the opcodes if they were
oldstyle when i should have been adding 1.  also, in my main loop, i was
checking to see if the opcode was HALT, but that was before i made the
correcting of the opcode for being old style.  the way i had it, an opcode
could be 26 in memory, and i would have interpreted that as a HALT command,
when it really should have been 27, an ADD command!