FIGURED OUT THE BUG WITH UNSAX
the command ``push 0" was what was messing the stack machine up.  as it was,
``push 0" would pop the top value off the stack and insert it into memory
location 0.  then, `SP' variable that the stack machine knows about would be
deposited into memory location 0.  i'm thinking of not even having Driver being
able to access the stack machine's SP, for simplicity.