works, but there is a memory leak
when the user loads up a file for the second time and tries to run SXX on that
program, the program either crashes or takes up an insane amount of clock
cycles.  i suspect this is because i'm creating a new swing worker thread when
there already exists one that's executing.  i'm going to have to think about
this one...